Add avatar uploading, and other new fields to profile page

+++ src/db/migrations/1_add_users_columns.php
@@ -7,14 +7,18 @@ class AddUsersColumns extends DatabaseBase {
         }
         
         $this->schema->table('users', function($table) {
-            $table->string('avatar');
-            $table->string('location');
-            $table->string('website');
-            $table->string('about');
+            $table->text('avatar')->nullable();
+            $table->string('location')->nullable();
+            $table->string('website')->nullable();
+            $table->string('about')->nullable();
         });
     }
 
     public function down() {
+        if ($this->schema->hasTable('users') == false) {
+            return;
+        }
+        
         $this->schema->table('users', function($table) {
             $table->dropColumn('avatar');
             $table->dropColumn('location');
